==> rmd/1_keywords.Rmd <==
---
title: "1_keywords"
output: html_document
---


```{r}
required_packages <- c("tidyverse", "readxl", "ggthemes", "hrbrthemes", "extrafont", "plotly", "scales", "stringr", "gganimate", "here", "tidytext", "sentimentr", "scales", "DT", "here", "sm", "mblm", "prettydoc", "reshape2", "treemapify", "glue", "magick", "imager", "fs", "knitr", "DataExplorer", "inspectdf", "rmdformats", "prettydoc", "janitor", "urltools")

for(i in required_packages) { 
if(!require(i, character.only = T)) {
#  if package is not existing, install then load the package
install.packages(i, dependencies = T)
require(i, character.only = T)
}
}
```

```{r}
#load city dataset
uscities_raw <- read_csv(here("raw_data", "uscities_raw.csv"))

#remove duplicate city name but keep with highest population
uscities_clean <- uscities_raw %>% 
  group_by(city) %>% #
  top_n(1, abs(population)) %>% 
  ungroup() 

# select certain properties and save new data set; to be used for mapping
us_city_details <- uscities_clean %>%
  # select(id, city, population, state_id, state_name, county_name, county_fips, lat, lng) %>% 
  arrange(desc(population)) %>% 
  mutate(city_id = row_number()) %>% 
  filter(population >= 100000) #cities larger or equal than 2000

```

```{r}
#select city_names 
us_city_names_id <- us_city_details %>% select(city_id, city)

#TO DO
#add keywords

us_city_names_id <- us_city_names_id %>% 
# mutate(car_accident_lawyer_end = str_glue("car accident lawyer {city}")) %>% 
# mutate(personal_injury_lawyer_end = str_glue("personal injury lawyer {city}")) %>% 
# mutate(personal_injury_attorney_end = str_glue("personal injury attorney {city}")) %>% 
# mutate(car_accident_attorney_end = str_glue("car accident attorney {city}")) %>%
mutate(car_accident_lawyer = str_glue("{city} car accident lawyer")) %>% 
mutate(personal_injury_lawyer = str_glue("{city} personal injury lawyer")) %>% 
mutate(car_accident_attorney = str_glue("{city} car accident attorney")) %>%   
mutate(personal_injury_attorney = str_glue("{city} personal injury attorney"))


#bind columns
final_keyword_list <- gather(us_city_names_id, key = "keyword_term", value = "keyword_strings", -city_id, -city)

final_keyword_list <- final_keyword_list %>% mutate(keyword = str_replace_all(final_keyword_list$keyword_term, "[[:punct:]]", " ")) %>% 
  select(-keyword_term)

#great overview on how to use gather(): https://uc-r.github.io/tidyr                     

```

```{r}
#save and export in csv to get keywords

write_csv(final_keyword_list, here("proc_data", "1_final_keyword_list.csv"))

write_csv(us_city_details, here("proc_data", "us_city_details_with_id.csv"))

```
