==> rmd/2_analysis.Rmd <==
---
title: "2_analysis"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r results=FALSE, message=FALSE, echo=FALSE}
save <- T
#create plots folder, if it doesn't exist
if(save && !dir.exists(here::here("plots"))) {
  dir.create(here::here("plots"))
  dir.create(here::here("plots", "png"))
  dir.create(here::here("plots", "csv"))
}

#load required packages, if package is not found try to install
required_packages <- c("xgboost", "data.table", "stringr", "SHAPforxgboost", "gridExtra", "wordcloud", "tm", "grDevices", "here", "pdftools", "plotly", "patchwork", "tidyverse", "ggwordcloud", "ggthemes")

not_installed_packages <- required_packages[!(required_packages %in% installed.packages()[,"Package"])]
if(length(not_installed_packages) > 0) 
  install.packages(not_installed_packages)
was_succesfully_loaded <- sapply(required_packages, require, character.only = TRUE)

if(any(!was_succesfully_loaded)) {
  cat(required_packages[!was_succesfully_loaded],"\n")
}
```

```{r plot-prep}
cols <- c("#d21c5c", "#060618")
caption <- "Source: Rankings.io"

## theme updates
theme_set(ggthemes::theme_clean(base_size = 18, base_family = "Alegreya Sans"))

theme_update(plot.background = element_rect(color = "white", 
                                            fill = "white"),
             panel.background = element_rect(color = NA, 
                                             fill = NA),
             plot.margin = margin(rep(20, 4)),
             plot.title = element_text(family = "Lora",
                                       color = "#060618", 
                                       size = 16, 
                                       hjust = .5, 
                                       margin = margin(b = 10)),
             plot.subtitle = element_text(#size = 9, 
                                          hjust = .5, 
                                          face = "plain", 
                                          color = "grey45", 
                                          margin = margin(b = 7)),
             plot.title.position = "plot",
             plot.caption = element_text(color = "grey40", size = 9, margin = margin(t = 12)),
             plot.caption.position = "plot",
             panel.grid.major.y = element_line(color = "grey80", linetype = "dotted", size = .4),
             axis.text = element_text(color = "grey45"),
             axis.ticks.length = unit(.15, "lines"),
             axis.line.x = element_line(color = "#060618"),
             axis.ticks.x = element_line(color = "#060618", 
                                         size = .3),
             axis.title.x = element_text(color = "#060618", 
                                         face = "bold", 
                                         margin = margin(t = 10)),
             axis.line.y = element_blank(), 
             axis.ticks.y = element_blank(),
             axis.title.y = element_text(color = "#060618", 
                                         face = "bold", 
                                         margin = margin(r = 10)),
             strip.background = element_rect(color = NA, 
                                             fill = NA))

## plotly function

plotly <- function(p, pos, w, h) {
  ggplotly(p, width = w*1.5, height = h*1.5) %>%
  config(displayModeBar = FALSE) %>%
  layout(
    margin = list(l = 50, r = 50, t = 100, b = pos*800),
    title = list(size = 13),
    annotations =
    list(x = 1, y = -pos, text = caption,
         showarrow = F, xref = 'paper', yref = 'paper',
         xanchor = 'right', yanchor = 'auto', xshift = 0, yshift = 0,
         font = list(size = 10, family = "Lora", color = "#666666"))
 )
}
```



## To-Do-List


## Data

Google My Business search results were collected for 4 unique keyword combinations in 426 US cities, giving in total 1659 search results. The format of the search queries was one the following:

* (city) + " car accident lawyer"  
* (city) + " personal injury lawyer" 
* (city) + " car accident attorney"
* (city) + " personal injury attorney"


#### Independent variables
All variables with prefix "Relative" are calculated as rank(values) / length(values) for values inside each search group. So for example, search with keyword "Milwaukee car accident lawyer" the entry with highest number of photos would get Relative number of photos value equal to 1, entry with lowest value would get 0, and the rest of the values would be something in between. The motivation behind this transformation is to make attribute values more comparable between searches, i.e. trying to minimize effect of the size of the population of the city etc.

* Attributes from basic.csv
  + Review rating
  + Relative # of reviews
  + Relative # of photos
  + Relative # of questions
  + Provides updates
  + Is business claimed
  + Has street address
  + Has website
  + Has phone
  + #Keywords in description
  + #Keywords in title
  + Has city in description
  + Has city in title 
  + Has lawyer or attorney in title
  + Has lawyer or attorney in description
  + Has car in description
  + Has car in title 
  + Has injury in description
  + Has injury in title
  + #Characters in description 
  + #Characters in title 
  + Type category is personal injury attorney 
  + Type category is lawyer 
  + Type category is law firm
  + Type category is empty
  + Type category is criminal justice attorney
  + Type category is other 

* Attributes from social_profile.csv
  + Has twitter
  + Has fb
  + Has youtube
  + Has instagram
  + Has linkedin
   
* Attributes from all_reviews.csv
  + Ratio of 5s in ratings 
  + Ratio of 4s in ratings
  + Ratio of 3s in ratings
  + Ratio of 2s in ratings
  + Ratio of 1s in ratings
  + Review response ratio
  + Relative # of review likes
  
* Attributes from q_a.csv
  + Question answer ratio
  
* Attributes from backlink_data_websites.csv
  + Relative domain_rating
  + Relative ahrefs_rank
  + Relative ref_domains_total
  + Relative ref_domains_dofollow
  + Relative linked_domains
  + Relative total_backlinks
  + Relative backlinks_no_follow
  + Relative total_keywords
  + Relative total_traffic

#### Dependent variables
* Position in the search results

```{r data, cache=TRUE, echo = FALSE}


#read csv
here::here("raw_data", "GMB data", "basic.csv")
basic <- fread(here::here("raw_data", "GMB data", "basic.csv"))
social_profile <- fread(here::here("raw_data", "GMB data", "social_profile.csv"))
user_review_featured <- fread(here::here("raw_data", "GMB data", "user_review_featured.csv"))
q_a <- fread(here::here("raw_data", "GMB data", "q_a.csv"))
all_reviews <- fread(here::here("raw_data", "GMB data", "all_reviews.csv"))
backlink_data_websites <- fread(here::here("raw_data", "GMB data", "backlink_data_websites.csv"))
cities <- fread(here::here("raw_data", "uscities_raw.csv"))
most_relevant_reviews <- fread(here::here("raw_data", "GMB data", "most_relevant_reviews.csv"))

#keep only the latest scraped rows for social_profile
social_profile[, scraping_date := as.Date(scraping_date)]
social_profile <- social_profile[, .SD[which.max(scraping_date)], by="place_id"]

#function to return relative values,
#ie. return new values between 0 and 1; zero for lowest and one for highest original value and the rest something between
relative_value <- function(values, na_as_zero = FALSE) {
  res <- (rank(values, na.last="keep", ties.method="average") - 1) / (sum(!is.na(values))-1)
  if(na_as_zero) {
    res[is.na(values)] <- 0
  }
  return(res)
}


#start processing data for modeling, create/transform attributes etc
data_all <- merge(basic, social_profile[, c("place_id","twitter","fb","youtube","instagram","linkedin"),with=F], by="place_id", all.x=T)
data_all[, target := relative_value(-position), by=keyword_strings]
data_all[, relative_position := relative_value(position), by=keyword_strings]

#remove result with google_ad =="Yes". These row are always in position 1 and 2 of the results
data_all <- data_all[google_ad == "No"]

#calculate search group sizes and remove a few searches without enough result
data_all[, group_size := .N, by=keyword_strings]
data_all <- data_all[group_size >= 1]

#relative variables, ie. normalize by dividing by median value of the whole search group
data_all[, 'Review rating' := ifelse(is.na(rating) | rating < 1 | rating > 5, NA_real_, rating), by=keyword_strings]
data_all[, 'Relative # of reviews' := relative_value(reviews, T), by=keyword_strings]
data_all[, 'Relative # of photos' := relative_value(number_of_photos, T), by=keyword_strings]
data_all[, 'Relative # of questions' := relative_value(total_questions, T), by=keyword_strings]

#transform Yes/No to 1/0
#data_all[, 'Has GoogleAd' := as.integer(google_ad == "Yes")]
data_all[, 'Provides updates' := as.integer(provides_updates == "Yes")]
data_all[, 'Is business claimed' := as.integer(business_claimed == "Yes")]

#does attribute have content or is it empty)
data_all[, 'Has street address' := as.integer(street_address != "")]
data_all[, 'Has website' := as.integer(website != "")]
data_all[, 'Has phone' := as.integer(phone != "")]
data_all[, 'Has twitter' := as.integer(twitter != "")]
data_all[, 'Has fb' := as.integer(fb != "")]
data_all[, 'Has youtube' := as.integer(youtube != "")]
data_all[, 'Has instagram' := as.integer(instagram != "")]
data_all[, 'Has linkedin' := as.integer(linkedin != "")]

#description & title
data_all[, '#Keywords in description' := sum(str_count(description, unlist(str_split(keyword_strings, " ")))) , by = row.names(data_all)]
data_all[, '#Keywords in title' := sum(str_count(title, unlist(str_split(keyword_strings, " ")))) , by = row.names(data_all)]
data_all[, 'Has city in description' := as.integer(grepl(city, description, ignore.case = TRUE)), by = row.names(data_all)]
data_all[, 'Has city in title' := as.integer(grepl(city, title, ignore.case = TRUE)), by = row.names(data_all)]
data_all[, 'Has lawyer or attorney in title' := as.integer(grepl("lawyer|attorney", title, ignore.case = TRUE)), by = row.names(data_all)]
data_all[, 'Has lawyer or attorney in description' := as.integer(grepl("lawyer|attorney", description, ignore.case = TRUE)), by = row.names(data_all)]
data_all[, 'Has car accident or personal injury in description' := as.integer(grepl("car accident|personal injury", description, ignore.case = TRUE)), by = row.names(data_all)]
data_all[, 'Has car accident or personal injury in title' := as.integer(grepl("car accident|personal injury", title, ignore.case = TRUE)), by = row.names(data_all)]
data_all[, '#Characters in description' := nchar(description)]
data_all[, '#Characters in title' := nchar(title)]

#type_category features
common_types <- c("Personal injury attorney",
                  "Lawyer",
                  "Law firm",
                  "",
                  "Criminal justice attorney" 
)
data_all[, 'Type category is personal injury attorney' := as.numeric(type_category == common_types[1])]
data_all[, 'Type category is lawyer' := as.numeric(type_category == common_types[2])]
data_all[, 'Type category is law firm' := as.numeric(type_category == common_types[3])]
data_all[, 'Type category is empty' := as.numeric(type_category == common_types[4])]
data_all[, 'Type category is criminal justice attorney' := as.numeric(type_category == common_types[5])]
data_all[, 'Type category is other' := as.numeric(!(type_category %in% common_types))]

data_all[, 'Has same city listed as in search query' := as.integer(city == city_place)]

#features from user_review_featured
data_all[, 'Has featured reviews' := as.integer(place_id %in% unique(user_review_featured$place_id))]

#features from all_reviews
all_reviews[, time_since_review := 0]
all_reviews[grepl("minute", all_reviews_date), time_since_review := 1]
all_reviews[grepl("hour", all_reviews_date), time_since_review := 2]
all_reviews[grepl("day", all_reviews_date), time_since_review := 3]
all_reviews[grepl("week", all_reviews_date), time_since_review := 4]
all_reviews[grepl("month", all_reviews_date), time_since_review := 5]
all_reviews[grepl("year", all_reviews_date), time_since_review := 6]

data_all <- merge(data_all, by="place_id", all.x=T, all.y=F,
      all_reviews[, list(
        'Ratio of 5s in ratings' = sum(all_reviews_rating == 5) / .N,
        'Ratio of 4s in ratings' = sum(all_reviews_rating == 4) / .N,
        'Ratio of 3s in ratings' = sum(all_reviews_rating == 3) / .N,
        'Ratio of 2s in ratings' = sum(all_reviews_rating == 2) / .N,
        'Ratio of 1s in ratings' = sum(all_reviews_rating == 1) / .N,
        'Review response ratio' = sum(all_reviews_response_from_owner != "") / .N,
        'review_median_of_likes' = as.numeric(median(all_reviews_number_of_likes, na.rm=TRUE)),
        'Time since previous review' = min(time_since_review)
      ), by = "place_id"]
)
data_all[, 'Relative # of review likes' := relative_value(review_median_of_likes, T), by="keyword_strings"]

#features from q_a
data_all <- merge(data_all, by="place_id", all.x=T, all.y=F,
                  q_a[, list(
                    'Question answer ratio' = sum(question_answer == "") / .N
                  ), by = "place_id"]
)

#features from cities
data_all[, city_population := 0.0]
data_all[, city_density := 0.0]
#find closest matching location in cities data given by Euclidean distance between lat&long
for(i in 1:nrow(data_all)) {
  pop <- NA_real_
  den <- NA_real_
  if(!is.na(data_all$latitude[i])) {
    city_index <- which.min((data_all$latitude[i] - cities$lat)^2 + (data_all$longtitude[i] - cities$lng)^2)
    pop <- cities$population[city_index]
    den <- cities$density[city_index]
  }
  data_all[i, city_population := pop]
  data_all[i, city_density := den]
 }
data_all[, 'Relative place population' := relative_value(city_population), by="keyword_strings"]


#features from most_relevant_reviews
data_all <- merge(data_all, by="place_id", all.x=T, all.y=F,
                  most_relevant_reviews[, list(
                    'Mean of most relevant ratings' = mean(most_relevant_reviews_rating == 5)
                  ), by = "place_id"]
)

#features from backlink_data_websites
data_all <- merge(data_all, backlink_data_websites, by="place_id", all.x=T, all.y=F)
data_all[, 'Relative domain_rating' := relative_value(domain_rating), by=keyword_strings]
data_all[, 'Relative ahrefs_rank' := relative_value(ahrefs_rank), by=keyword_strings]
data_all[, 'Relative ref_domains_total' := relative_value(ref_domains_total), by=keyword_strings]
data_all[, 'Relative ref_domains_dofollow' := relative_value(ref_domains_dofollow), by=keyword_strings]
data_all[, 'Relative linked_domains' := relative_value(linked_domains), by=keyword_strings]
data_all[, 'Relative total_backlinks' := relative_value(total_backlinks), by=keyword_strings]
data_all[, 'Relative backlinks_no_follow' := relative_value(backlinks_no_follow), by=keyword_strings]
data_all[, 'Relative total_keywords' := relative_value(total_keywords), by=keyword_strings]
data_all[, 'Relative total_traffic' := relative_value(total_traffic), by=keyword_strings]

```


## Modelling

The goal of the statistical modelling in this study is to find answers to three key questions:

1. How accurately can the rankings be predicted given the dependent variables?
2. What are the most important features for the predictions?
3. What is the direction of the impact?

The method of choice for the study was gradient boosted decision trees (GBDT). GBDT is a widely used machine learning technique which can be used in many settings from regression or classification to learning to rank type of problems. In a learning to rank problem, there is a ordered list of items and the goal for the model is to calculate a score for each item based on dependent variables such that the original order is retained. 

In process of building the model, data set was split to two folds: train data (containing around 70% of searches) and test data (the rest of the data, about 30%). GBDT model was fitted using training data, predictions were calculated for the test data set, and then finally predictions were compared to real observed rankings. The chosen evaluation metric was Spearman's rank correlation coefficient. Spearman's rank correlation is a scaled measurement of the agreement of two rankings. Perfectly matching rankings would give value of 1, the expected value for random rankings is zero and reverse order would have value of -1.

The next step is to understand why the model makes particular predictions; what are the most important dependent variables and how their values effect the predictions? For this purpose SHapley Additive exPlanations (SHAP) values were calculated. In SHAP each prediction is presented as a sum of each dependent variable's responsibility. Then the overall impact of any particular variable can be measured as a average of absolute values over the whole data set.


```{r model, cache=TRUE, echo = FALSE}
#for xgboost rank modeling, rows have to be ordered by group
setorder(data_all, keyword_strings, position)

#split data to train and test sets based on search term
set.seed(1)
train_ratio <- 0.7
ids <- unique(data_all$keyword_strings)
train_ids <- sample(ids, floor(train_ratio*length(ids)))
test_ids <- setdiff(ids, train_ids)

feature_cols <- c(
'Review rating',
'Relative # of reviews',
'Relative # of photos',
'Relative # of questions',

#'Has GoogleAd',
'Provides updates',
'Is business claimed',

'Has street address',
'Has website',
'Has phone',
'Has twitter',
'Has fb',
'Has youtube',
'Has instagram',
'Has linkedin',

'#Keywords in description',
'#Keywords in title',
'Has city in description',
'Has city in title' ,
'Has lawyer or attorney in title',
'Has lawyer or attorney in description',
'Has car accident or personal injury in description',
'Has car accident or personal injury in title' ,
'#Characters in description', 
'#Characters in title', 

'Type category is personal injury attorney', 
'Type category is lawyer', 
'Type category is law firm',
'Type category is empty',
'Type category is criminal justice attorney',
'Type category is other', 

'Has same city listed as in search query',

'Has featured reviews', 

'Ratio of 5s in ratings', 
'Ratio of 4s in ratings',
'Ratio of 3s in ratings',
'Ratio of 2s in ratings',
'Ratio of 1s in ratings',
'Review response ratio',
'Time since previous review',

'Relative # of review likes',

'Question answer ratio',

'Relative domain_rating',
'Relative ahrefs_rank',
'Relative ref_domains_total',
'Relative ref_domains_dofollow',
'Relative linked_domains',
'Relative total_backlinks',
'Relative backlinks_no_follow',
'Relative total_keywords',
'Relative total_traffic',

'Relative place population',
'Mean of most relevant ratings')

data_train <- data_all[keyword_strings %in% train_ids, ]
X_train <- as.matrix(data_train[, feature_cols, with=F])
y_train <- data_all[keyword_strings %in% train_ids]$target
group_train <- unique(data_all[keyword_strings %in% train_ids,c("keyword_strings", "group_size")])$group_size

data_test <- data_all[keyword_strings %in% test_ids, ]
X_test <- as.matrix(data_test[, feature_cols, with=F])
y_test <- data_all[keyword_strings %in% test_ids]$target
group_test <- unique(data_all[keyword_strings %in% test_ids,c("keyword_strings", "group_size")])$group_size


#model settings
dtrain = xgb.DMatrix(data = X_train, label = y_train, group = group_train)
dtest = xgb.DMatrix(data = X_test, label = y_test, group = group_test)

params <- list(eta = 0.1, 
               subsample = 0.7, 
               colsample_bytree = 0.7, 
               min_child_weight = 1,
               max_leaves = 64,
               max_depth = 0,
               lambda = 0,
               alpha = 0.5,
               grow_policy = "lossguide",
               tree_method = "hist",
               objective = "rank:pairwise"
          )
nrounds <- 1000

#model training
model <- xgb.train(data = dtrain, 
                 params = params, 
                 nrounds = nrounds, 
                 verbose = 0, 
                 watchlist = list(train=dtrain, test=dtest))

#predict
pred <- predict(model, X_test)

#shap
shap_values <- shap.values(xgb_model = model, X_train = X_train)
shap_importance <- data.frame(feature = names(shap_values$mean_shap_score), 
                              importance = shap_values$mean_shap_score,
                              stringsAsFactors = F)

shap_values_test <- shap.values(model, X_test)
shap_scores_test <- shap_values_test$shap_score

#spearman correlations
spearman_data <- data_all[keyword_strings %in% test_ids,]
spearman_data[, pred := pred]
spearman_data <- spearman_data[, list(correlation=cor(target, pred, method="spearman")), by="keyword_strings"]
```

## Results

#### Accuracy

* In the plot below, showing the distribution of correlations calculated separately for each search
* Overall mean correlation is about 0.6, showing fairly good fit between observations and predictions

```{r accuracy, echo = FALSE}
mean_correlation <- format(round(mean(spearman_data$correlation), 2), nsmall = 2)
spearman_plot <- ggplot(spearman_data, aes(x = correlation)) + 
  geom_bar(fill=cols[1], color="white") +
  xlab("Correlations for search rankings") +
  ylab("Count") + 
  scale_x_binned() +
  scale_y_continuous(expand = c(0,0), breaks = 0:5 * 20) +
  labs(title = paste0("Mean Spearman's correlations between predicted and observed rankings: ", mean_correlation), caption = caption)
#plot(spearman_plot)

if(save == T){ 
  ggsave(here::here("plots", "spearman.pdf"), spearman_plot, 
         width = 10, height = 7, device = cairo_pdf) 

  # interactive chart pushed to plotly chart studio
  g <- plotly(spearman_plot, .2, w = 1000, h = 700)
  api_create(g, filename = "spearman_plot")
}
```

### Feature importance

* The first plot is showing feature importance, that is, each feature's average contribution to model's predictions. 
  
* Second plot shows the direction of the impact given feature's value. 

* Overall, the most interesting features for us to look at are a) high in the first plot and b) show clear pattern in the second plot. 

* E.g. if we look at the first row and the feature named "Has same city listed as in search query", we can see kind of polarized distribution of SHAP values around zero. Yellow points correspond to high feature values (in this case, "No") and here their impact to all predictions in the data set is negative and therefore making model belief that predicted positions should be worse for them. Where as purple points correspond to high feature values ("Yes") and have positive impact for predicted positions.

* Similarly, "Type category is personal injury" and "Type category is personal injury" are similar too the "Has same city listed as in search query". Ie. if they have value "Yes" they will impact positively to predicted positions.

```{r importance, echo = FALSE}

feature_importance_plot <- ggplot(shap_importance, aes(x=reorder(feature, importance), y=importance, fill=importance)) +
  geom_bar(stat='identity') + 
  scale_fill_gradient2(low = cols[2], mid = cols[2], high = cols[1], midpoint = 0) + 
  xlab("Feature") +
  ylab("Importance") + 
  labs(title = "When a city is used in a search query,\nthe city in a GMB profile is the dominant ranking factor") + 
  theme(legend.position="none", 
        plot.title.position = "plot",
        #axis.text.x=element_blank(), 
        #axis.ticks.x=element_blank(),
        axis.text.x = element_text(size = 9, color = "grey45"),
        #axis.title.x = element_text(size = 11, color = "grey45", margin = margin(t = 5)),
        axis.text.y = element_text(size = 10, color = "black"),
        panel.grid.major.x = element_line(color = "grey70", linetype = "dotted", size = .25),
        panel.grid.major.y = element_blank()) +  
  scale_y_continuous(limits=c(0.0,NA), expand=c(0,0))+
  coord_flip() +
  labs(x = NULL, caption = caption) +
  theme(plot.title = element_text(margin = margin(b = 15)))
#plot(feature_importance_plot)


shap_values_plot <- shap.plot.summary.wrap2(shap_values$shap_score, X_train, dilute=3)
#plot(shap_values_plot)

if(save == T){ 
  ggsave(here::here("plots", "feature_importance.pdf"), feature_importance_plot, 
         width = 7.5, height = 10, device = cairo_pdf) 
  #ggsave(here::here("plots", "shap_values.pdf"), shap_values_plot, 
  #       width = 12, height = 10.5, device = cairo_pdf) 
  
  # interactive chart pushed to plotly chart studio
  g <- plotly(feature_importance_plot, .15, w = 1000, h = 1500)
  api_create(g, filename = "feature_importance_plot")
}

```


## A closer look at the features

The depended variables used in this study can be roughly organized into five main groups, these are listed below and also showing a few important variables suggested by SHAP values.

* Location
  + Has same city listed as in search query
  + Relative place population

* Popularity
  + Relative # of photos 
  + Relative # of questions

* Type category
  + Type category is personal injury attorney/lawyer

* Keywords and title/description
  + Has "lawyer or attorney"/city in title
  + #Characters in description

* Reviews
  + Relative # of reviews
  + Review rating

In terms of SEO, the first two categories are not much of a interest as they are something difficult or even impossible to change or adjust, but the last three are more interesting and worth further investigation. 

```{r features, results='asis', echo=FALSE}
#data for all visualizations
#to make it cleaner to make plots, keep only searches with exactly 20 results. 
#Ie. keep only searches with >= 20 results and then if more than 20, remove the extra rows
#keeps about 90% of all rows
data_vis <- data_all[group_size >= 20,]
data_vis <- data_vis[, .SD[order(position) <= 20], by="keyword_strings"]
data_vis <- data_vis[, position := rank(position, ties.method="first"), by="keyword_strings"]

#helper function for printing
to_percentage <- function(x, digits=2) {paste0(round(100*x, digits=digits), "%")}

```  
### Type category
```{r type_category_table, results='asis', echo=FALSE}

#gather basic info
type_counts <- as.data.frame(sort(table(data_vis[type_category!=""]$type_category), decreasing=T), stringsAsFactors = F)
names(type_counts)[1] <- "category"
type_unique_data <- data_vis[, list(N = length(unique(type_category))), by="keyword_strings"]

#build table
rs <- list()
rs[[length(rs)+1]] <- list('Total unique categories', 
                           length(unique(data_all$type_category)) - 1)

rs[[length(rs)+1]] <- list('Missing type category', 
                         to_percentage(sum(data_all$type_category=="") / nrow(data_all)))

rs[[length(rs)+1]] <- list('Categories with more than >=10 results', 
                         sum(type_counts$Freq>=10))

rs[[length(rs)+1]] <- list('Categories with more than >=100 results', 
                           sum(type_counts$Freq>=100))

rs[[length(rs)+1]] <- list('Categories with more than >=1000 results', 
                           sum(type_counts$Freq>=1000))

rs[[length(rs)+1]] <- list('Median unique categories in one search', 
                           median(type_unique_data$N))

rs[[length(rs)+1]] <- list('Min unique categories in one search', 
                           min(type_unique_data$N))

rs[[length(rs)+1]] <- list('Max unique categories in one search', 
                           max(type_unique_data$N))

type_table <- rbindlist(rs)
setnames(type_table, 1:2, c("Type", "Value"))

knitr::kable(type_table, caption = "Basic information about type categories")

if(save == T){ 
  write.csv(type_table, here::here("plots", "csv", "type_table.csv"), row.names = F, quote=F)
}
``` 

```{r type_category_plots, results='asis', echo=FALSE}
#plots
keep_top_N <- 10
type_freq_plot <- 
  type_counts[1:keep_top_N,] %>% 
  mutate(perc = round(Freq / sum(Freq) * 100)) %>% 
  ggplot(aes(x=reorder(category, Freq), y=Freq)) +
  geom_bar(stat="identity", fill=cols[1]) +
  geom_text(aes(label = glue::glue("{format(Freq, big.mark = ",")} ({perc}%)")), 
            family = "Lora", fontface = "bold", color = "grey45", 
            lineheight = .9, nudge_y = 750, hjust = 0) +
  ylab("Count") +
  xlab("Category") + 
  labs(title = "Ten most common type categories in search results") +
  scale_y_continuous(expand = c(0, 0), limits = c(0, 27000)) +
  theme(plot.title = element_text(hjust = 0.5), 
        axis.text.y = element_text(color = "black"), 
        panel.grid.major.y = element_blank()) +
  coord_flip() +
  labs(x = NULL, caption = caption)
#plot(type_freq_plot)

title <- "Personal injury attorney outranks other common business types"

ps <- list()
for(i in 1:6) {
  ps[[i]] <- ggplot(data_vis[type_category == type_counts$category[i]], aes(x=position)) +
  geom_bar(fill=cols[1], width = .9) +
  scale_x_continuous(breaks = c(1, 5, 10, 15, 20), expand = c(.02, .02)) +
  labs(title = type_counts$category[i], x = "Position", y = NULL) + 
  theme(plot.title = element_text(face = "plain", size = 14), 
        axis.title = element_text(size = 11, face = "plain"))
}

type_position_plot <- 
  ps[[1]] + ps[[2]] + ps[[3]] + ps[[4]] + ps[[5]] + ps[[6]] +
  plot_layout(nrow = 2) +
  plot_annotation(title = title, caption = caption)
#type_position_plot <- grid.arrange(ps[[1]], ps[[2]], ps[[3]],
#             ps[[4]], ps[[5]], ps[[6]], nrow=2, ncol=3, top = "Distribution of search positions for 6 most common type categories")
#plot(type_position_plot)

## for plotly
variable <- type_counts$category[1:6]

plot_facet <- function(p) {
  ggplotly(p + theme(axis.text = element_text(size = 8))) %>% 
    add_annotations(
      text = p$labels$title,
      x = 0,
      y = 1,
      yref = "paper",
      xref = "paper",
      xanchor = "left",
      yanchor = "top",
      showarrow = FALSE,
      font = list(size = 11)
    )
}

type_position_plotly <- 
  subplot(
    plot_facet(ps[[1]]),
    plot_facet(ps[[2]]),
    plot_facet(ps[[3]]),
    plot_facet(ps[[4]]),
    plot_facet(ps[[5]]),
    plot_facet(ps[[6]])
  ) %>% 
  layout(
    title = title, 
    titlefont = list(size = 22, family = "Lora")
  )

## same y axis scaling for all
ps <- list()
for(i in 1:6) {
  ps[[i]] <- ggplot(data_vis[type_category == type_counts$category[i]], aes(x=position)) +
  geom_bar(fill=cols[1], width = .9) +
  scale_x_continuous(breaks = c(1, 5, 10, 15, 20), expand = c(.02, .02)) +
  scale_y_continuous(expand = c(0, 0), limits = c(0, 1600), breaks = seq(0, 1500, by = 250)) +
  labs(title = type_counts$category[i], x = "Position", y = NULL) + 
  theme(plot.title = element_text(face = "plain", size = 14, margin = margin(b = 15)), 
        axis.title = element_text(size = 11, face = "plain"))
}

type_position_plot_scaled <- 
  ps[[1]] + ps[[2]] + ps[[3]] + ps[[4]] + ps[[5]] + ps[[6]] +
  plot_layout(nrow = 2) +
  plot_annotation(title = title, caption = caption)
#type_position_plot <- grid.arrange(ps[[1]], ps[[2]], ps[[3]],
#             ps[[4]], ps[[5]], ps[[6]], nrow=2, ncol=3, top = "Distribution of search positions for 6 most common type categories")
#plot(type_position_plot_scaled)

## for plotly
type_position_plotly_scaled <- 
  subplot(
    plot_facet(ps[[1]]),
    plot_facet(ps[[2]]),
    plot_facet(ps[[3]]),
    plot_facet(ps[[4]]),
    plot_facet(ps[[5]]),
    plot_facet(ps[[6]])
  )


if(save == T){ 
  ggsave(here::here("plots", "type_freq.pdf"), type_freq_plot + theme(axis.text.y = element_text(size = 16)), 
         width = 12.5, height = 8, device = cairo_pdf) 
  ggsave(here::here("plots", "type_position.pdf"), type_position_plot, 
         width = 12.5, height = 8, device = cairo_pdf) 
  ggsave(here::here("plots", "type_position_scaled.pdf"), type_position_plot_scaled, 
         width = 12.5, height = 8, device = cairo_pdf) 
  
  ## !! Plotly doesnt work with multipanel plots!
  # interactive chart pushed to plotly chart studio
  g <- plotly(type_freq_plot, .15, w = 1800, h = 900)
  api_create(g, filename = "type_freq_plot")
  g <- plotly(type_position_plotly, .12, w = 2250, h = 800) %>% 
        layout(
          title = title, 
          titlefont = list(size = 18, family = "Lora")
        )
  api_create(g, filename = "type_position_plot")
  g <- plotly(type_position_plotly_scaled, .12, w = 2250, h = 800) %>% 
        layout(
          title = title, 
          titlefont = list(size = 18, family = "Lora")
        )
  api_create(g, filename = "type_position_plot_scaled")
}
```    

Key takeaways:

* Distributions for more general categories (Lawyer, Law firm, Legal services) are tilted towards lower positions in the results compared to best matching category (Personal injury attorney).
* Same thing occurs with specific, but less matching categories (Criminal justice attorney, Family law attorney)
* However, some of these categories (Law firm, Criminal justice attorney, Legal services) have relatively large counts for top positions. One possible explanation is that these groups represent larger companies with high revenues and therefore would be ranked higher. 

### Titles and descriptions
```{r title_description_table, results='asis', echo=FALSE, message=FALSE}


#gather basic info about titles & descriptions

rs <- list()
rs[[length(rs)+1]] <- list("Median character length (non missing)", 
                           median(data_all[title!=""]$`#Characters in title`),
                           median(data_all[description!=""]$`#Characters in description`))

rs[[length(rs)+1]] <- list("Min character length (non missing)", 
                           min(data_all[title!=""]$`#Characters in title`),
                           min(data_all[description!=""]$`#Characters in description`))

rs[[length(rs)+1]] <- list("Max character length (non missing)",
                           max(data_all[title!=""]$`#Characters in title`),
                           max(data_all[description!=""]$`#Characters in description`))

rs[[length(rs)+1]] <- list('Missing', 
                          to_percentage(sum(data_all$title=="") / nrow(data_all)),
                          to_percentage(sum(data_all$description=="") / nrow(data_all)))

rs[[length(rs)+1]] <- list("Containing lawyer or attorney", 
                           to_percentage(sum(data_all$`Has lawyer or attorney in title`==1) / nrow(data_all)),
                          to_percentage(sum(data_all$`Has lawyer or attorney in description`==1) / nrow(data_all)))

rs[[length(rs)+1]] <- list("Containing car accident or personal injury", 
                          to_percentage(sum(data_all$`Has car accident or personal injury in title`==1) / nrow(data_all)),
                          to_percentage(sum(data_all$`Has car accident or personal injury in description`==1) / nrow(data_all)))

rs[[length(rs)+1]] <- list("Containing city name", 
                           to_percentage(sum(data_all$`Has city in title`==1) / nrow(data_all)),
                           to_percentage(sum(data_all$`Has city in description`==1) / nrow(data_all)))

title_description_table <- rbindlist(rs)
setnames(title_description_table, 1:3, c("Type", "Title", "Description"))

knitr::kable(title_description_table, caption = "Basic information about titles and descriptions")

if(save == T){ 
  write.csv(title_description_table, here::here("plots", "csv", "title_description_table.csv"), row.names = F, quote=F)
}
```

```{r title_description_wordclouds, results='asis', echo=FALSE, message=FALSE}
generate_wordcloud_data <- function(texts) {

  #get words as a list of vectors
  word_list <- lapply(texts, function(x) unlist(strsplit(x, " ")))
  
  #remove duplicates
  word_list <- word_list[!duplicated(lapply(word_list, sort))]
  
  #clean up words
  docs <- Corpus(VectorSource(word_list))
  docs <- tm_map(docs, removeNumbers)
  docs <- tm_map(docs, removePunctuation)
  docs <- tm_map(docs, stripWhitespace)
  docs <- tm_map(docs, content_transformer(tolower))
  docs <- tm_map(docs, removeWords, stopwords("english"))
  
  #calculate dtm & freqs
  dtm <- TermDocumentMatrix(docs) 
  matrix <- as.matrix(dtm) 
  words <- sort(rowSums(matrix),decreasing=TRUE) 
  df <- data.frame(word = names(words),freq=words, stringsAsFactors = F)
  
  return(list(dtm = matrix, words = df))
}

#disable warnings
defaultW <- getOption("warn")
options(warn = -1)
old_par <- par()

words_titles <- generate_wordcloud_data(data_all[title!=""]$title)
words_desc <- generate_wordcloud_data(data_all[description!=""]$description)

set.seed(1)

desc_wordcloud <-
  words_desc$words %>% 
  as_tibble %>% 
  top_n(100, freq) %>%  
  mutate(word = str_to_upper(word),
         angle = 90 * sample(c(0, 1), n(), replace = TRUE, prob = c(60, 40))) %>%
  ggplot(aes(label = word, size = freq, color = freq)) +
    geom_text_wordcloud(
      #aes(angle = angle),
      family = "Alegreya",
      fontface = "bold",
      shape = "circle",
      grid_margin = 2.5
    ) +
    scale_color_gradient(low = "grey75", high = cols[1]) +
    scale_size_area(max_size = 10) +
    labs(title = "\n100 most common words in the description", caption = caption) +
    theme_minimal() +
    theme(plot.title = element_text(hjust = .5, family = "Lora", face = "bold"),
          plot.caption = element_text(hjust = 1, family = "Alegreya", size = 7, color = "grey40"))
#plot(desc_wordcloud)

title_wordcloud <- 
  words_titles$words %>% 
  as_tibble %>% 
  top_n(100, freq) %>%  
  mutate(word = str_to_upper(word),
         angle = 90 * sample(c(0, 1), n(), replace = TRUE, prob = c(60, 40))) %>%
  ggplot(aes(label = word, size = freq, color = freq)) +
    geom_text_wordcloud(
      #aes(angle = angle),
      family = "Alegreya",
      fontface = "bold",
      shape = "circle",
      grid_margin = 2.5
    ) +
    scale_color_gradient(low = "grey75", high = cols[1]) +
    scale_size_area(max_size = 18) +
    labs(title = "\n100 most common words in the title", caption = caption) +
    theme_minimal() +
    theme(plot.title = element_text(hjust = .5, family = "Lora", face = "bold"),
          plot.caption = element_text(hjust = 1, family = "Alegreya", size = 7, color = "grey40"))
#plot(title_wordcloud)


if(save) {
  ggsave(here::here("plots", "descriptions_wordcloud.pdf"), desc_wordcloud, width=8, height=6, device = cairo_pdf)
  ggsave(here::here("plots", "titles_wordcloud.pdf"), title_wordcloud, width=8, height=4.5, device = cairo_pdf)
}

# if(save) {
#   pdf(here::here("plots", "descriptions_wordcloud.pdf"), width=8, height=8)
# }
# layout(matrix(c(1, 2), nrow=2), heights=c(1, 4))
# par(mar=rep(0, 4))
# plot.new()
# text(x=0.5, y=0.5, "Descriptions word cloud")
# descriptions_wordcloud_plot <- wordcloud(words = words_desc$words$word, 
#                                          freq = words_desc$words$freq, min.freq = 3,
#                                          max.words=100, random.order=FALSE, rot.per=0.35, 
#                                          colors=c(colorspace::lighten(cols[1], .7), cols[1]))#brewer.pal(8, "Dark2"))
# if(save) {
#   dev.off()
# }
# 
# if(save == T) {
#   pdf(here::here("plots", "titles_wordcloud.pdf"), width=8, height=8)
# }
# layout(matrix(c(1, 2), nrow=2), heights=c(1, 4))
# par(mar=rep(0, 4))
# plot.new()
# text(x=0.5, y=0.5, "Titles word cloud")
# titles_wordcloud_plot <- wordcloud(words = words_titles$words$word, 
#                                    freq = words_titles$words$freq, min.freq = 3,
#                                    max.words=100, random.order=FALSE, rot.per=0.35,
#                                    colors=c(colorspace::lighten(cols[1], .5), cols[1]))#brewer.pal(8, "Dark2"))
# 
# if(save == T) {
#   dev.off()
# }
# 
# #back to default
# par(old_par) 
# options(warn = defaultW)

```

```{r title_description_plots, results='asis', echo=FALSE, message=FALSE}
plot_count_title <- ggplot(data_vis["title"!=""], aes(y=`#Characters in title`, x=as.factor(position))) +
  geom_boxplot(color=cols[1], fill="grey85") +
  xlab("Position") +
  ylab(NULL) +
  labs(title = "Title")
#plot(plot_count_title)
plot_count_description <-  ggplot(data_vis[description!=""], aes(y=`#Characters in description`, x=as.factor(position))) +
  geom_boxplot(color=cols[1], fill="grey85") +
  xlab("Position") +
  ylab("Number of Characters") +
  labs(title = "Description")
#plot(plot_count_description)

#title_description_counts_plot <- grid.arrange(plot_count_title, plot_count_description, nrow=1, ncol=2, top="Non-empty title and description length boxplots")

title_description_counts_plot <- 
  (plot_count_description + plot_count_title) * theme(plot.title = element_text(face = "plain", size = 14)) +
  plot_annotation(title = "Non-empty title and description length boxplots")
#plot(title_description_counts_plot)

## !! subplot keeps only last ttiel so no title per subplot possible (?)
#subplot(ggplotly(plot_count_title), ggplotly(plot_count_description))


# plot_missing_description <- ggplot(data_vis[description==""], aes(x=position)) +
#   geom_histogram(color=cols[1], fill="grey85", bins=20) +
#   ylab("Count") +
#   xlab("Position") + 
#   labs(title = "Yes") +
#   theme(plot.title = element_text(hjust = 0.5)) + facet_wrap()
# plot_not_missing_description <- ggplot(data_vis[description!=""], aes(x=position)) +
#   geom_histogram(color=cols[1], fill="grey85", bins=20) +
#   ylab("Count") +
#   xlab("Position") + 
#   labs(title = "No") +
#   theme(plot.title = element_text(hjust = 0.5))
# description_missing_plot <- grid.arrange(plot_missing_description, plot_not_missing_description, nrow=1, ncol=2, top="Empty description")

description_missing_plot_facet <- 
  data_vis %>% 
  mutate(cat = if_else(description == "", "Empty", "Non-empty"),
         cat = factor(cat, levels = c("Empty", "Non-empty"))) %>% 
  ggplot(aes(x=position)) +
  geom_bar(fill=cols[1], width = .8) +
  scale_x_continuous(breaks = c(1, 5, 10, 15, 20), expand = c(.02, .02)) +
  scale_y_continuous(expand = c(0,0))  +
  ylab("Count") +
  xlab(NULL) + 
  labs(title = "Non-empty description in GMB profile help to increase position", caption = caption) + 
  facet_wrap(~cat)
#plot(description_missing_plot_facet)

# plot_car_personal_title0 <- ggplot(data_vis[`Has car accident or personal injury in title` == 0], aes(x=position)) +
#   geom_histogram(color="white", fill=cols[1], bins=20) +
#   ylab("Count") +
#   xlab("Position") + 
#   labs(title = "No \"car accident\" or \"personal injury\"") +
#   theme(plot.title = element_text(hjust = 0.5))
# plot_car_personal_title1 <- ggplot(data_vis[`Has car accident or personal injury in title` == 1], aes(x=position)) +
#   geom_histogram(color="white", fill=cols[1], bins=20) +
#   ylab("Count") +
#   xlab("Position") + 
#   labs(title = "Has \"car accident\" or \"personal injury\"") +
#   theme(plot.title = element_text(hjust = 0.5))

title_t <- "Including relevant keywords in GMB profile title helps to increase position"
title_d <- "Including relevant keywords in GMB profile description helps to increase position"

description_car_personal_plot <-
  data_vis %>% 
  mutate(cat = if_else(`Has car accident or personal injury in description` == 1, "\"Car accident\" or \"personal injury\" in description", "\"Car accident\" or \"personal injury\" not in description"),
         cat = factor(cat, levels = c("\"Car accident\" or \"personal injury\" in description", "\"Car accident\" or \"personal injury\" not in description"))) %>% 
  ggplot(aes(x=position)) +
  geom_bar(fill=cols[1], width = .8) +
  scale_x_continuous(breaks = c(1, 5, 10, 15, 20), expand = c(.02, .02)) +
  scale_y_continuous(expand = c(0,0))  +
  labs(x = NULL, y = "", title = title_d, caption = caption) +
  facet_wrap(~cat)
#plot(description_car_personal_plot)


# plot_car_personal_desc0 <- ggplot(data_vis[`Has car accident or personal injury in description` == 0], aes(x=position)) +
#   geom_histogram(color="white", fill=cols[1], bins=20) +
#   ylab("Count") +
#   xlab("Position") + 
#   labs(title = "No \"car accident\" or \"personal injury\"") +
#   theme(plot.title = element_text(hjust = 0.5))
# plot_car_personal_desc1 <- ggplot(data_vis[`Has car accident or personal injury in description` == 1], aes(x=position)) +
#    geom_histogram(color="white", fill=cols[1], bins=20) +
#    ylab("Count") +
#    xlab("Position") + 
#    labs(title = "Has \"car accident\" or \"personal injury\"") +
#    theme(plot.title = element_text(hjust = 0.5))

title_car_personal_plot <-
  data_vis %>% 
  mutate(cat = if_else(`Has car accident or personal injury in title` == 1, "\"Car accident\" or \"personal injury\" in title", "\"Car accident\" or \"personal injury\" not in title"),
         cat = factor(cat, levels = c("\"Car accident\" or \"personal injury\" in title", "\"Car accident\" or \"personal injury\" not in title"))) %>% 
  ggplot(aes(x=position)) +
  geom_bar(fill=cols[1], width = .8) +
  scale_x_continuous(breaks = c(1, 5, 10, 15, 20), expand = c(.02, .02)) +
  scale_y_continuous(expand = c(0,0))  +
  ylab("Count") +
  xlab("Position") + 
  labs(x = NULL, y = "", title = title_t, cpation = caption) +
  facet_wrap(~cat)
#plot(title_car_personal_plot)


# plot_lawyer_attorney_desc0 <- ggplot(data_vis[`Has lawyer or attorney in description` == 0], aes(x=position)) +
#   geom_histogram(color="white", fill=cols[1], bins=20) +
#   ylab("Count") +
#   xlab("Position") + 
#   labs(title = "No \"lawyer\" or \"attorney\"") +
#   theme(plot.title = element_text(hjust = 0.5))
# plot_lawyer_attorney_desc1 <- ggplot(data_vis[`Has lawyer or attorney in description` == 1], aes(x=position)) +
#   geom_histogram(color="white", fill=cols[1], bins=20) +
#   ylab("Count") +
#   xlab("Position") + 
#   labs(title = "Has \"lawyer\" or \"attorney\"") +
#   theme(plot.title = element_text(hjust = 0.5))

description_lawyer_attorney_plot <-
  data_vis %>% 
  mutate(cat = if_else(`Has lawyer or attorney in description` == 1, "\"Lawyer\" or \"attorney\" in description", "\"Lawyer\" or \"attorney\" not in description"),
         cat = factor(cat, levels = c("\"Lawyer\" or \"attorney\" in description", "\"Lawyer\" or \"attorney\" not in description"))) %>% 
  ggplot(aes(x=position)) +
  geom_bar(fill=cols[1], width = .8) +
  scale_x_continuous(breaks = c(1, 5, 10, 15, 20), expand = c(.02, .02)) +
  scale_y_continuous(expand = c(0,0))  +
  labs(x = NULL,  y = "", title = title_d, caption = caption) +
  facet_wrap(~cat)
#plot(description_lawyer_attorney_plot)

# plot_lawyer_attorney_title0 <- ggplot(data_vis[`Has lawyer or attorney in title` == 0], aes(x=position)) +
#   geom_histogram(color="white", fill=cols[1], bins=20) +
#   ylab("Count") +
#   xlab("Position") + 
#   labs(title = "No \"lawyer\" or \"attorney\"") +
#   theme(plot.title = element_text(hjust = 0.5))
# plot_lawyer_attorney_title1 <- ggplot(data_vis[`Has lawyer or attorney in title` == 1], aes(x=position)) +
#   geom_histogram(color="white", fill=cols[1], bins=20) +
#   ylab("Count") +
#   xlab("Position") + 
#   labs(title = "Has \"lawyer\" or \"attorney\"") +
#   theme(plot.title = element_text(hjust = 0.5))

title_lawyer_attorney_plot <-
  data_vis %>% 
  mutate(cat = if_else(`Has lawyer or attorney in title` == 1, "\"Lawyer\" or \"attorney\" in title", "\"Lawyer\" or \"attorney\" not in title"),
         cat = factor(cat, levels = c("\"Lawyer\" or \"attorney\" in title", "\"Lawyer\" or \"attorney\" not in title"))) %>% 
  ggplot(aes(x=position)) +
  geom_bar(fill=cols[1], width = .8) +
  scale_x_continuous(breaks = c(1, 5, 10, 15, 20), expand = c(.02, .02)) +
  scale_y_continuous(expand = c(0,0))  +
  labs(x = NULL, y = "", title = title_t, caption = caption) +
  facet_wrap(~cat)
#plot(title_lawyer_attorney_plot)


# plot_city_desc0 <- ggplot(data_vis[`Has same city listed as in search query`==1], aes(x=position)) +
#   geom_histogram(color="white", fill=cols[1], bins=20) +
#   ylab("Count") +
#   xlab("Position") + 
#   labs(title = "No city name") +
#   theme(plot.title = element_text(hjust = 0.5))
# plot_city_desc1 <- ggplot(data_vis[`Has city in description` == 1], aes(x=position)) +
#   geom_histogram(color="white", fill=cols[1], bins=20) +
#   ylab("Count") +
#   xlab("Position") + 
#   labs(title = "Has city name") +
#   theme(plot.title = element_text(hjust = 0.5))

description_city_plot <-
  data_vis %>% 
  mutate(cat = if_else(`Has same city listed as in search query` == 1, "City name in description", "City name not in description"),
         cat = factor(cat, levels = c("City name in description", "City name not in description"))) %>% 
  ggplot(aes(x=position)) +
  geom_bar(fill=cols[1], width = .8) +
  scale_x_continuous(breaks = c(1, 5, 10, 15, 20), expand = c(.02, .02)) +
  scale_y_continuous(expand = c(0,0)) +
  labs(x = NULL, y = "", title = "Including the city name in GMB profile description helps to increase position", caption = caption) +
  theme() +
  facet_wrap(~cat)
#plot(description_city_plot)

# plot_city_title0 <- ggplot(data_vis[`Has city in title` == 0], aes(x=position)) +
#   geom_histogram(color="white", fill=cols[1], bins=20) +
#   ylab("Count") +
#   xlab("Position") + 
#   labs(title = "No city name") +
#   theme(plot.title = element_text(hjust = 0.5))
# plot_city_title1 <- ggplot(data_vis[`Has city in title` == 1], aes(x=position)) +
#   geom_histogram(color="white", fill=cols[1], bins=20) +
#   ylab("Count") +
#   xlab("Position") + 
#   labs(title = "Has city name") +
#   theme(plot.title = element_text(hjust = 0.5))

title_city_plot <-
  data_vis %>% 
  mutate(cat = if_else(`Has city in title` == 1, "City name in title", "City name not in title"),
         cat = factor(cat, levels = c("City name in title", "City name not in title"))) %>% 
  ggplot(aes(x=position)) +
  geom_bar(fill=cols[1], width = .8) +
  scale_x_continuous(breaks = c(1, 5, 10, 15, 20), expand = c(.02, .02)) +
  scale_y_continuous(expand = c(0,0))  +
  labs(x = NULL, y = "", title = "Including the city name in GMB profile title helps to increase position", caption = caption) +
  facet_wrap(~cat)
#plot(title_city_plot)

# title_keywords_plot <- grid.arrange(plot_car_personal_title0, plot_car_personal_title1,
#              plot_lawyer_attorney_title0, plot_lawyer_attorney_title1,
#              plot_city_title0, plot_city_title1,
#              nrow=3, ncol=2, top="Title containing keywords")
# 
# description_keywords_plot <- grid.arrange(plot_car_personal_desc0, plot_car_personal_desc1,
#              plot_lawyer_attorney_desc0,  plot_lawyer_attorney_desc1,
#              plot_city_desc0, plot_city_desc1,
#              nrow=3, ncol=2, top="Description containing keywords")

if(save == T){ 
  ggsave(here::here("plots", "title_description_counts.pdf"), title_description_counts_plot, 
         width = 10, height = 6, device = cairo_pdf) 
  #ggsave(here::here("plots", "description_missing.pdf"), description_missing_plot, 
  #       width = 12.5, height = 8, device = cairo_pdf) 
  ggsave(here::here("plots", "description_missing.pdf"), description_missing_plot_facet, 
         width = 10, height = 5.5, device = cairo_pdf) 
  # ggsave(here::here("plots", "title_keywords.pdf"), title_keywords_plot, 
  #        width = 12.5, height = 8, device = cairo_pdf) 
  # ggsave(here::here("plots", "description_keywords.pdf"), description_keywords_plot, 
  #        width = 12.5, height = 8, device = cairo_pdf)
  ggsave(here::here("plots", "description_car_personal.pdf"), description_car_personal_plot,
         width = 10, height = 5.5, device = cairo_pdf)
  ggsave(here::here("plots", "title_car_personal.pdf"), title_car_personal_plot,
         width = 10, height = 5.5, device = cairo_pdf)
  ggsave(here::here("plots", "description_lawyer_attorney.pdf"), description_lawyer_attorney_plot,
         width = 10, height = 5.5, device = cairo_pdf)
  ggsave(here::here("plots", "title_lawyer_attorney.pdf"), title_lawyer_attorney_plot,
         width = 10, height = 5.5, device = cairo_pdf)
  ggsave(here::here("plots", "description_city.pdf"), description_city_plot,
         width = 10, height = 5.5, device = cairo_pdf)
  ggsave(here::here("plots", "title_city.pdf"), title_city_plot,
         width = 10, height = 5.5, device = cairo_pdf)
  
  # interactive chart pushed to plotly chart studio
  g <- plotly(plot_count_title + ggtitle("Non-empty title length boxplots"), .15, w = 900, h = 600)
  api_create(g, filename = "plot_count_title")
  g <- plotly(plot_count_description + ggtitle("Non-empty description length boxplots"), .15, w = 900, h = 600)
  api_create(g, filename = "plot_count_description")
  g <- plotly(description_missing_plot_facet, .15, w = 1200, h = 450)
  api_create(g, filename = "description_missing_plot_facet")
  g <- plotly(description_car_personal_plot, .12, w = 1200, h = 450)
  api_create(g, filename = "description_car_personal_plot")
  g <- plotly(title_car_personal_plot, .12, w = 1200, h = 450)
  api_create(g, filename = "title_car_personal_plot")
  g <- plotly(description_lawyer_attorney_plot, .12, w = 1200, h = 450)
  api_create(g, filename = "description_lawyer_attorney_plot")
  g <- plotly(title_lawyer_attorney_plot, .12, w = 1200, h = 450)
  api_create(g, filename = "title_lawyer_attorney_plot")
  g <- plotly(description_city_plot, .12, w = 1200, h = 450)
  api_create(g, filename = "description_city_plot")
  g <- plotly(title_city_plot, .12, w = 1200, h = 450)
  api_create(g, filename = "title_city_plot")
}
```  

Key takeaways:

* In general, length of titles/descriptions don't correlate much with positions
* The only exception is businesses with missing description; they tend to have clearly lower positions in results
* In all cases with different keywords, search results containing keyword in title/description have lower positions on average than entries without
* The effect is more noticeable for titles than descriptions
* Also, more specific words (both city name and "car accident"/"personal injury") have higher effect then lawyer/attorney

### Reviews
```{r reviews_table, results='asis', echo=FALSE}

rs <- list()
rs[[length(rs)+1]] <- list("Median #reviews", 
                           median(data_all$reviews, na.rm=T))

rs[[length(rs)+1]] <- list("Max #reviews", 
                           max(data_all$reviews, na.rm=T))

rs[[length(rs)+1]] <- list("No reviews available", 
                           to_percentage(sum(is.na(data_all$rating)) / nrow(data_all)))

rs[[length(rs)+1]] <- list("Average rating", 
                           round(mean(data_all$rating, na.rm=T), digits=2))

rs[[length(rs)+1]] <- list("Response ratio by owners", 
                          to_percentage(sum(all_reviews$all_reviews_response_from_owner!="")/nrow(all_reviews)))

rs[[length(rs)+1]] <- list("Average number of likes per review", 
                           round(mean(all_reviews$all_reviews_number_of_likes, na.rm=T), digits=2))

rating_table <- rbindlist(rs)
setnames(rating_table, 1:2, c("Type", "Value"))

knitr::kable(rating_table, caption = "Basic information about reviews")

if(save == T){ 
  write.csv(rating_table, here::here("plots", "csv", "rating_table.csv"), row.names = F, quote=F)
}
```  

```{r reviews_plots, results='asis', echo=FALSE}

rating_distribution_plot <- 
  all_reviews %>% 
  as_tibble() %>% 
  dplyr::mutate(total = n(), all_reviews_rating = factor(all_reviews_rating, levels = 1:5)) %>% 
  dplyr::group_by(all_reviews_rating) %>% 
  dplyr::mutate(n = n(), perc = round(n / total * 100)) %>% 
  ## summarise does soemhow not work with data.table....
  #dplyr::summarise(perc = total / n) %>% 
  dplyr::slice(1) %>% 
  ggplot(aes(x=all_reviews_rating, y=n)) +
  geom_col(fill=cols[1]) +
  geom_text(aes(label = glue::glue("{format(n, big.mark = ',')} ({perc}%)")), 
            family = "Lora", fontface = "bold", color = "grey45", size = 3.5,
            hjust = .5, lineheight = .9, nudge_y = 45000) +
  scale_y_continuous(labels = scales::comma_format(), expand = c(0, 0), limits = c(0, 1200000), breaks = seq(0, 1000000, by = 250000)) +
  labs(x = "Ratings", y = "Count", title = "Distribution of ratings", caption = caption)
#plot(rating_distribution_plot)

review_position_plot <- ggplot(data_vis, aes(y=`Relative # of reviews`, x=as.factor(position))) +
  geom_boxplot(color=cols[1], fill="grey85") +
  labs(x = "Position", y = "Relative # of reviews", title = "Higher review volumes correlates with higher GMB rank ", caption = caption)
#plot(review_position_plot)

rating_position_plot <- ggplot(data_vis[!is.na(`Review rating`)], aes(y=`Review rating`, x=as.factor(position))) +
  geom_boxplot(color=cols[1], fill="grey85") +
  labs(x = "Position", y = "Review rating", title = "Review rating per position", caption = caption) 
#plot(rating_position_plot)

if(save == T){ 
  ggsave(here::here("plots", "rating_distribution.pdf"), rating_distribution_plot, 
         width = 9, height = 6, device = cairo_pdf) 
  ggsave(here::here("plots", "review_position.pdf"), review_position_plot, 
         width = 9, height = 6, device = cairo_pdf) 
  ggsave(here::here("plots", "rating_position.pdf"), rating_position_plot, 
         width = 9, height = 6, device = cairo_pdf) 
  
  # interactive chart pushed to plotly chart studio
  g <- plotly(rating_distribution_plot, .2, w = 1000, h = 600)
  api_create(g, filename = "rating_distribution_plot")
  g <- plotly(review_position_plot, .2, w = 1000, h = 500)
  api_create(g, filename = "review_position_plot")
  g <- plotly(rating_position_plot, .2, w = 1000, h = 500)
  api_create(g, filename = "rating_position_plot")
}
```  

Key takeaways:

* Businesses with highest number of reviews tend to have the lowest positions (top left corner)
* In contrast, low number of reviews correlates with higher position (bottom right)
* Perhaps surprisingly, ratings themself don't seem to practically any effect on page's position, only activity matters. 
* But then on the other hand, if almost 90% of ratings are five stars and average rating is over 4.5, maybe there just isn't much room for a differentiation.


### Backlink

```{r ref_domains_dofollow_table, results='asis', echo=FALSE}

rs <- list()
rs[[length(rs)+1]] <- list('Median', 
                           median(data_all$ref_domains_dofollow, na.rm=T),
                           median(data_all$total_traffic, na.rm=T),
                           median(data_all$ahrefs_rank, na.rm=T),
                           median(data_all$domain_rating, na.rm=T))

rs[[length(rs)+1]] <- list('Min', 
                           min(data_all$ref_domains_dofollow, na.rm=T),
                           min(data_all$total_traffic, na.rm=T),
                           min(data_all$ahrefs_rank, na.rm=T),
                           min(data_all$domain_rating, na.rm=T))

rs[[length(rs)+1]] <- list('Max', 
                           max(data_all$ref_domains_dofollow, na.rm=T),
                           max(data_all$total_traffic, na.rm=T),
                           max(data_all$ahrefs_rank, na.rm=T),
                           max(data_all$domain_rating, na.rm=T))

rs[[length(rs)+1]] <- list("Missing", 
                           to_percentage(sum(is.na(data_all$ref_domains_dofollow)) / nrow(all_reviews)),
                           to_percentage(sum(is.na(data_all$total_traffic)) / nrow(all_reviews)),
                           to_percentage(sum(is.na(data_all$ahrefs_rank)) / nrow(all_reviews)),
                           to_percentage(sum(is.na(data_all$domain_rating)) / nrow(all_reviews)))

backlink_table <- rbindlist(rs)
setnames(backlink_table, 1:5, c("Type", "ref_domains_dofollow", "total_traffic", "ahrefs_rank", "domain_rating"))

knitr::kable(backlink_table, caption = "Basic information about backlinks")

if(save == T){ 
  write.csv(backlink_table, here::here("plots", "csv", "backlink_table.csv"), row.names = F, quote=F)
}
```

```{r ref_domains_dofollow_plots, results='asis', echo=FALSE}
ref_domains_dofollow_plot <- ggplot(data_vis[!is.na(`Relative ref_domains_dofollow`)],
       aes(y=`Relative ref_domains_dofollow`, x=as.factor(position))) +
  geom_boxplot(color=cols[1], fill="grey85") +
  ylab("Relative number of do-follow domains") +
  xlab("Position") +
  labs(title = "Referring do-follow domains appear to slightly impact GMB rank", caption = caption) 
#plot(ref_domains_dofollow_plot)

total_traffic_plot <- ggplot(data_vis[!is.na(`Relative total_traffic`)],
       aes(y=`Relative total_traffic`, x=as.factor(position))) +
  geom_boxplot(color=cols[1], fill="grey85") +
  ylab("Relative total_traffic") +
  xlab("Position") +
  labs(title = "Relative total_traffic boxplot", caption = caption) 
#plot(total_traffic_plot)

ahrefs_rank_plot <- ggplot(data_vis[!is.na(`Relative ahrefs_rank`)],
       aes(y=`Relative ahrefs_rank`, x=as.factor(position))) +
  geom_boxplot(color=cols[1], fill="grey85") +
  ylab("Ahrefs ranking") +
  xlab("Position") +
  labs(title = "High Ahrefs rank correlates with higher GMB position", caption = caption)
#plot(ahrefs_rank_plot)

domain_rating_plot <- ggplot(data_vis[!is.na(`Relative domain_rating`)],
       aes(y=`Relative domain_rating`, x=as.factor(position))) +
  geom_boxplot(color=cols[1], fill="grey85") +
  ylab("Relative domain_rating") +
  xlab("Position") +
  labs(title = "Relative domain_rating boxplot", caption = caption)
#plot(domain_rating_plot)

if(save == T){ 
  ggsave(here::here("plots", "ref_domains_dofollow.pdf"), ref_domains_dofollow_plot, 
         width = 9, height = 6, device = cairo_pdf) 
  ggsave(here::here("plots", "total_traffic.pdf"), total_traffic_plot, 
         width = 9, height = 6, device = cairo_pdf)
  ggsave(here::here("plots", "ahrefs_rank.pdf"), ahrefs_rank_plot, 
         width = 9, height = 6, device = cairo_pdf)
  ggsave(here::here("plots", "domain_rating.pdf"), domain_rating_plot, 
         width = 9, height = 6, device = cairo_pdf)
  
  # interactive chart pushed to plotly chart studio
  g <- plotly(ref_domains_dofollow_plot, .2, w = 1000, h = 550)
  api_create(g, filenmae = "ref_domains_dofollow_plot")
  g <- plotly(total_traffic_plot, .2, w = 1000, h = 550)
  api_create(g, filename = "total_traffic_plot")
  g <- plotly(ahrefs_rank_plot, .2, w = 1000, h = 550)
  api_create(g, filename = "ahrefs_rank_plot")
  g <- plotly(domain_rating_plot, .2, w = 1000, h = 550)
  api_create(g, filename = "domain_rating_plot")
}
```  

Key takeaways:

* Higher ref_domains_dofollow and total_traffic are more common for better positions
* ahrefs_rank seems to be telling the same story, the only difference is that small number for rank is better, so the shape is inverted



```{r other_features_table, results='asis', echo=FALSE}
rs <- list()
rs[[length(rs)+1]] <- list('Median #photos', 
                           median(data_all$number_of_photos, na.rm=T))
rs[[length(rs)+1]] <- list('Max #photos', 
                           max(data_all$number_of_photos, na.rm=T))
rs[[length(rs)+1]] <- list('Zero #photos', 
                           to_percentage(sum(is.na(data_all$number_of_photos)) / nrow(data_all)))
rs[[length(rs)+1]] <- list('Provides Google updates', 
                           to_percentage(sum(data_all$provides_updates == "Yes") / nrow(data_all)))

misc_features_table <- rbindlist(rs)
setnames(misc_features_table, 1:2, c("Type", "Value"))

knitr::kable(misc_features_table, caption = "Basic information about #photos and Google updates")

if(save == T){ 
  write.csv(misc_features_table, here::here("plots", "csv", "misc_features_table.csv"), row.names = F, quote=F)
}
```

### Provided updates and #photos

```{r other_features_plot, results='asis', echo=FALSE}

number_of_photos_plot <- ggplot(data_vis, aes(y=`Relative # of photos`, x=as.factor(position))) +
  geom_boxplot(color=cols[1], fill="grey85") +
  labs(x = "Position", y = "Relative number of photos", title = "Higher review volumes correlates with higher GMB rank", caption = caption)
#plot(number_of_photos_plot)

# updates0 <- ggplot(data_vis[`Provides updates` == 0], aes(x=position)) +
#     geom_histogram(color="white", fill=cols[1], bins=20) +
#     ylab("Count") +
#     xlab("Position") + 
#     labs(title = "No") +
#     theme(plot.title = element_text(hjust = 0.5))
# 
# updates1 <- ggplot(data_vis[`Provides updates` == 1], aes(x=position)) +
#     geom_histogram(color="white", fill=cols[1], bins=20) +
#     ylab("Count") +
#     xlab("Position") + 
#     labs(title = "Yes") +
#     theme(plot.title = element_text(hjust = 0.5))
# 
# provides_updates_plot <- grid.arrange(updates1, updates0, nrow=1, ncol=2, top="Provides updates on Google")

provides_updates_plot_facet <- 
  data_vis %>% 
  mutate(cat = if_else(`Provides updates` == 1, "Yes", "No"),
         cat = factor(cat, levels = c("Yes", "No"))) %>% 
  ggplot(aes(x=position)) +
  geom_bar(fill=cols[1], width = .8) +
  scale_x_continuous(breaks = c(1, 5, 10, 15, 20), expand = c(.02, .02)) +
  scale_y_continuous(expand = c(0,0))  +
  labs(x= "\nPosition", y = "Count", title = "Provides updates on Google", caption = caption) + 
  facet_wrap(~cat)
#plot(provides_updates_plot_facet)


if(save == T){ 
  ggsave(here::here("plots", "number_of_photos.pdf"), number_of_photos_plot, 
         width = 9, height = 6, device = cairo_pdf) 
  ggsave(here::here("plots", "provides_updates.pdf"), provides_updates_plot_facet, 
         width = 9, height = 6, device = cairo_pdf) 
  
  # interactive chart pushed to plotly chart studio
  g <- plotly(number_of_photos_plot, .2, w = 1000, h = 550)
  api_create(g, filename = "number_of_photos_plot")
  g <- plotly(provides_updates_plot_facet, .2, w = 1000, h = 600)
  api_create(g, filename = "provides_updates_plot_facet")
}
``` 

Key takeaways:

* Businesses who provide Google updates more often tend to have better positions.
* Similarly, also higher number of photos is correlated with better positions.
* So once again, activity seems to be the key



```{r data_table, include=FALSE}

rs <- list()
rs[[length(rs)+1]] <- list('#Cities', 
                           length(unique(data_all$city)))
rs[[length(rs)+1]] <- list('#Keywords', 
                           length(unique(data_all$keyword)))
rs[[length(rs)+1]] <- list('Total # of searches', 
                           length(unique(data_all$keyword_strings)))
rs[[length(rs)+1]] <- list('#Unique place IDs', 
                           length(unique(data_all$place_id)))
rs[[length(rs)+1]] <- list('Searches with less than 10 results', 
                           to_percentage(sum(data_all$group_size<10) / nrow(data_all)))
rs[[length(rs)+1]] <- list('Searches with 10 to 15 results', 
                           to_percentage(sum(data_all$group_size>=10 &data_all$group_size<=15) / nrow(data_all)))
rs[[length(rs)+1]] <- list('Searches with 16 to 19 results', 
                           to_percentage(sum(data_all$group_size>=16 &data_all$group_size<=19) / nrow(data_all)))
rs[[length(rs)+1]] <- list('Searches with a full first-page', 
                           to_percentage(sum(data_all$group_size>=20) / nrow(data_all)))

basic_info_table <- rbindlist(rs)
setnames(basic_info_table, 1:2, c("Type", "Value"))

knitr::kable(basic_info_table, caption = "Basic information about the data set")

if(save == T){ 
  write.csv(basic_info_table, here::here("plots", "csv", "basic_info_table.csv"), row.names = F, quote=F)
}

```


```{r convert_to_pngs, results='asis', echo=FALSE}
if(save == T) {
  pdfs <- list.files(here::here("plots"), pattern = "*.pdf")
  
  setwd(here::here("plots", "png"))
  for(pdf in pdfs) {
    pdf_convert(pdf = here::here("plots", pdf), format = "png", dpi = 720)
  }
}
```
