==> rmd/3_final.Rmd <==
---
title: "Google My Business Ranking Study"
author: 
- "Lead: Chris Dreyer ([rankings.io](https://rankings.io/))"
- "Support: FrontPage Data ([frontpagedata.com](https://frontpagedata.com/))"
date: "Last updated on `r Sys.Date()`"
output: 
  html_document:
    theme: paper
    highlight: kate
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: false
editor_options:
  chunk_output_type: console
---

<style>
.list-group-item.active, .list-group-item.active:hover, .list-group-item.active:focus {
  background-color: #D21C5C;
    border-color: #D21C5C;
}

body {
  font-family: montserrat;
  color: #444444;
    font-size: 14px;
}

h1 {
  font-weight: bold;
  font-size: 28px;
}

h1.title {
  font-size: 30px;
  color: #D21C5C;
}

h2 {
  font-size: 24px;
}

h3 {
  font-size: 18px;
}
</style>
  
# 1 Introduction

Search Results containing a local pack often get the majority of clicks. Knowing which local rankings factors to optimize for the biggest bang is crucial for SEOs and business owners alike.

Smaller data studies as well as opinion-based surveys have sought to uncover the relevance and importance of local ranking factors, including but not limited to the [localseoguide.com](https://www.localseoguide.com/guides/local-seo-ranking-factors/), [moz.com](https://moz.com/local-search-ranking-factors) and [brigtlocal.com](https://www.brightlocal.com/research/how-car-dealerships-are-speeding-ahead-with-google-my-business/).

However, in our view, most of the studies are outdated and contain severe statistical and methodological flaws.

This study intends to fill the gap and shed some insights in the personal injury niche on which local ranking factors are the most relevant ones. 

It builds and extents on our previous [data-based study to evaluate 112,000 personal injury law SERPs (search engine results page)](https://rankings.io/personal-injury-seo-data-science/). You can see a full breakdown of the previous data analysis right [here](https://rankings.io/data-report/). 


# 2 Methodology

##### Step 1 Keyword Selection

We defined for 4 unique keyword combinations in 426 US cities (> 100.000 inihabitants). The format of the search queries was the following:

* (city) + "car accident lawyer"
* (city) + "personal injury lawyer"
* (city) + "car accident attorney"
* (city) + "personal injury attorney"

##### Step 2 Data Mining

To gather the base data for the study, we created a script to collect data points from the Google My Business Maps Listing. The relevant entries were scraped from the Google Search page (https://www.google.com/) by entering the above keyword combinations. A basic data overview can be found below. 

It is important to note that we did not collect at data on proximity factors, given the inherent and practical difficulties in obtaning such data.   

```{r data_table, echo=FALSE}
  basic_info_table <- read.csv(here::here("plots", "csv", "basic_info_table.csv"), stringsAsFactors=F)
  knitr::kable(basic_info_table, caption = "Basic information about the data set")
```

##### Step 3 Data Enrichment

As a next step, we enriched the listed website domains with SEO data frm the third-data provider Ahrefs. To do so, we cut down the URL website domains to their root and uploaded them onto the Ahrefs bulk analysis tool. All data sets were then merged into one. 


##### Step 4 Data Analysis

We applied a state of the art machine learning model (first published in 2017) to determine the importance of GMB factors on rankings. More information on the model can be found in the Technical Annex. Then, we provided a deep dive into single variables that the model identified as particulary important to impact GMB positions. 


# 3 Research Findings 


## 3.1 The importance of individial ranking factors 

The plot below indicates what factors are particulry important in impacting GMB rankings (a more technical explanation can be found below). We can conclude that having the same GMB city listed as in the search query has the largest effect on the ranking position, followed by the "type category is personal injury lawyer", # of reviews and the # of photos. Adding the string "lawyer or "attorney" to the title can also positively impact positions, according to our analysis. 

GMB details such as adding street address, website domain, and phone number do not seem to be relevant. The same is true for social signals. 



```{r, out.extra='style="border: none;"', out.width="100%", frameborder="0", echo=FALSE}
knitr::include_url("https://plotly.com/~dancoup/168.embed?link=false&amp;modebar=false&amp;autosize=true", height="800")
```


##### Technical explanation 

The shap feature importance plot (see above) indicates the importance of each variable. That is each factor's average contribution to the model's predictions. The higher a variable is listed on the plot, the higher the factor´s contribution is to the GMB rankings.

On the other hand, the plot below shows the direction of the impact given each factor's value.

For instance, if we look at the first row and the feature named "Has same city listed as in search query", we can see a polarized distribution of SHAP values around zero. Yellow points correspond to low feature values (in this case, "No"). That means that their impact to all predictions in the data set is negative. The purple points correspond to high feature values ("Yes") and have a positive impact on the predicted positions.

To take another example, the "Type category is personal injury" variable behaves similarly to the "Has same city listed as in search query" in that sense that they have higher feature values i.e. they will impact positively the predicted positions.


```{r shap_values, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "shap_values_1.png"))
```

##### Accuracy

The plot below shows the distribution of correlations between model's predictions and observed positions calculated separately for each Google search.
Overall, the mean correlation is about 0.6, showing fairly good fit between observations and predictions.

```{r spearman, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "spearman_1.png"))
```


## 3.2 A closer look at individual ranking factors

The depended variables used in this study can be roughly organized into five main groups. These are listed below and also showing a few important variables suggested by the model results. 

* Type category
  + Type category is personal injury attorney/lawyer

* Keywords and title/description
  + Has "lawyer or attorney"/city in title
  + Number of characters in description
  + Has same city listed as in search query
  
* Reviews
  + Number of reviews
  + Review ratings
  
* SEO
  + Number of referring domains (dofollow)
  + Ahrefs rank
  + total traffic 
  + domain rating

* Activity
  + Number of photos 
  + Provides updates on Google

### 3.2.1 Type category

```{r type_table, echo=FALSE}
  type_table <- read.csv(here::here("plots", "csv", "type_table.csv"), stringsAsFactors=F)
  knitr::kable(type_table, caption = "Basic information about type categories")
```

```{r type_freq, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "type_freq_1.png"))
```

```{r type_position, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "type_position_1.png"))
```

###### {#common-business-types}

```{r, out.extra='style="border: none;"', out.width="100%", frameborder="0", echo=FALSE}
knitr::include_url("https://plotly.com/~dancoup/216.embed?link=false&amp;modebar=false&amp;autosize=true", height="600")
```



Key takeaways:

* Distributions for more general categories ("Lawyer", "Law firm", "Legal services") are tilted towards lower positions in the search results compared to the best matching category ("Personal injury attorney").
* Same trend occurs with specific, but less matching categories ("Criminal justice attorney", "Family law attorney")
* However, some of these categories ("Law firm", "Criminal justice attorney", "Legal services") have relatively large counts for top positions. 



### 3.2.2 Title and description
```{r title_description_table, echo=FALSE}
  title_description_table <- read.csv(here::here("plots", "csv", "title_description_table.csv"), stringsAsFactors=F)
knitr::kable(title_description_table, caption = "Basic information about titles and descriptions")
```

```{r titles_wordcloud, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "titles_wordcloud_1.png"))
```

```{r descriptions_wordcloud, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "descriptions_wordcloud_1.png"))
```

```{r title_description_counts, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "title_description_counts_1.png"))
```

```{r description_missing, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "description_missing_1.png"))
```


###### {#lawyer-attorney-title}

```{r, out.extra='style="border: none;"', out.width="100%", echo=FALSE}
knitr::include_url("https://plotly.com/~dancoup/186.embed?link=false&amp;modebar=false&amp;autosize=true", height="600")
```

###### {#lawyer-attorney-description}
```{r, out.extra='style="border: none;"', out.width="100%", echo=FALSE}
knitr::include_url("https://plotly.com/~dancoup/184.embed?link=false&amp;modebar=false&amp;autosize=true", height="600")
```


###### {#personal-car-accident-title}

```{r, out.extra='style="border: none;"', out.width="100%", echo=FALSE}
knitr::include_url("https://plotly.com/~dancoup/182.embed?link=false&amp;modebar=false&amp;autosize=true", height="600")
```

###### {#personal-car-accident-description}

```{r, out.extra='style="border: none;"', out.width="100%", frameborder="0", echo=FALSE}
knitr::include_url("https://plotly.com/~dancoup/182.embed?link=false&amp;modebar=false&amp;autosize=true", height="600")
```


```{r title_keywords_city, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "title_city_1.png"))
```

```{r description_keywords_city, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "description_city_1.png"))
```

Key takeaways:

* In general, length of titles/descriptions doesn't correlate with positions. 
* The only notable exception are businesses with missing descriptions; they tend to have lower positions in the search results. 
* GMB listings containing various keyword combinations in the title/description exhibit, on average, higher positions than entries without.
* The effect is more noticeable for titles than descriptions. 
* In addition, more specific words (both city names and "car accident"/"personal injury") have higher effects then keywords such as "lawyer" or "attorney".  



### 3.2.3 Reviews
```{r reviews_table, echo=FALSE}
  rating_table <- read.csv(here::here("plots", "csv", "rating_table.csv"), stringsAsFactors=F)
  knitr::kable(rating_table, caption = "Basic information about reviews")
```

```{r rating_distribution, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "rating_distribution_1.png"))
```

```{r rating_position, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "rating_position_1.png"))
```

```{r review_position, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "review_position_1.png"))
```

Key takeaways:

* GMB listings with higher number of reviews tend to have  higher positions (top left corner).
* In contrast, low number of reviews correlate with lower positions (bottom right).
* Perhaps surprisingly, review ratings themself don't seem to show any effect on GMB listing´s position. It seems that only review activity matters. However, if almost 90% of ratings are five star ratings with an average rating of over 4.5, there perhaps isn't much room for a differentiation.

### 3.2.4 SEO
```{r backlink_table, echo=FALSE}
  backlink_table <- read.csv(here::here("plots", "csv", "backlink_table.csv"), stringsAsFactors=F)
  knitr::kable(backlink_table, caption = "Basic information about backlinks")
```
```{r ref_domains_dofollow, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "ref_domains_dofollow_1.png"))
```

```{r total_traffic, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "total_traffic_1.png"))
```

```{r ahrefs_rank, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "ahrefs_rank_1.png"))
```

```{r domain_rating, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "domain_rating_1.png"))
```

Key takeaways:

* Higher # of referring do-follow domains (ref_domains_dofollow), total traffic numbers as well as domain rating are positivey related with higher GMB positions.
* The Ahrefs rank seems to be telling the same story. The only difference is that a lower ahrefs rank number seems to be indicate higher positions, so the shape is inverted. 


### 3.2.5 Provided updates and number of photos

```{r other_features_table, echo=FALSE}
  misc_features_table <- read.csv(here::here("plots", "csv", "misc_features_table.csv"), stringsAsFactors=F)
  knitr::kable(misc_features_table, caption = "Basic information about #photos and Google updates")
```

```{r provides_updates, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "provides_updates_1.png"))
```
```{r number_of_photos, echo=FALSE}
  knitr::include_graphics(here::here("plots", "png", "number_of_photos_1.png"))
```    
Key takeaways:

* GMB listings which provide more frequent Google updates tend to indicate higher positions.
* Similarly, the number of photos is positively correlated with better GMB positions.
* So once again, activity seems to be the key. 


# 4 Technical Annex 

The goal of the statistical model in this study is to find answers to three key questions:

1. How accurately can the rankings be predicted given the dependent variables?
2. What are the most important features for the predictions?
3. What is the direction of the impact?

The statistical method of choice for this study was the gradient boosted decision trees (GBDT) model. The GBDT is a widely used machine learning technique which can be used in many settings. These can range from regression and classification to learning to rank type of problems. In a learning to rank problem, there is a ordered list of items and the goal for the model is to calculate a score for each item based on the dependent variables such that the original order is retained. 

In process of building the model, data set was split into two pieces: the training data set (containing around 70% of searches) and testing data set (the remainder with about 30%). The GBDT model was fitted using training data, predictions were calculated for the test data set. Then the predictions were compared to real observed rankings. The chosen evaluation metric was Spearman's rank correlation coefficient. Spearman's rank correlation is a scaled measurement of the agreement of two rankings. Perfectly matching rankings would provide a value of 1, the expected value for random rankings is zero. The reverse order would have value of -1.

The next step is to understand why the model makes particular predictions; what are the most important dependent variables and how their values effect the predictions? For this purpose, the SHapley Additive exPlanations (SHAP) values were calculated. In SHAP, each prediction is presented as a sum of each dependent variable's responsibility. Then the overall impact of any particular variable can be measured as a average of absolute values over the whole data set.

**Note on the variables**

All variables with the prefix “Relative” are calculated as rank (values) / length (values) for values inside each search group. For example, a search compromised of “Milwaukee car accident lawyer”, the entry with the highest number of photos would get the relative # of photos value equal to 1. The entry with the lowest value would get a 0, and the remaining values would be something in between. The motivation behind this transformation is to make attribute values more comparable between search results i.e. trying to minimize effect of the size of the population of the city. 

Some of the variables included in the model can be considered to be control variables. That means that they are in itself nof of interest of the analysis, but have to be included to get more accurate views of impact of other factors. One of the control varaibles is place population.
